Rename assignment base64 helpers module to core/base64
