+++ oot3dhdtextgenerator/core/base64.py
